feat(#2454): Slice 13 GREEN — migrate team(action=create) to typed in-process runtime service
Extract ~300 lines of CREATE_TEAM logic from api/handlers/team.rs into
a new neutral typed module (team_ops.rs) with a CreateTeamRequest struct.
Both API and MCP handlers become thin adapters that parse transport-specific
input and delegate to team_ops::create(). The MCP path no longer uses
crate::api::call loopback — it receives registry/notifier via RuntimeContext.

- New src/team_ops.rs: CreateTeamRequest, plan_member_names, build_member_entries, create()
- api/handlers/team.rs: thin adapter parsing params → CreateTeamRequest
- mcp/handlers/task.rs: runtime-aware adapter, structured error when runtime=None
- mcp/handlers/dispatch.rs: threads ctx.runtime to handle_create_team
- Budget reduced: production api::call sites 5→4


Agend-Agent: archfix-opus-3
Agend-Task: t-20260719051207831029-80319-3
Agend-Branch: red/2454-create-team-runtime
Agend-Issued-At: 2026-07-19T10:58:27.124467+00:00
